// src/models/models/steam/steam-items/ItemStickerModel.ts
export default interface StickerOnItemModel{
    
    /**
     * The name of the sticker
     * @type {string}
     * @memberof ItemStickerModel
     */
    name: string;


    /**
     * The url for the sticker image
     * @type {string}
     * @memberof ItemStickerModel
     */
    imageUrl: string;
}